base_laser_mount
This is two versions of a simple laser mount.

version 1 has a square hole for the usb cable, version 2 has the side
with the usb cable shorter, so no hole needed.
Because I do not know what type of 90degree mini usb we've ordered.

This is for @mas-group/mas_industrial_robotics#59