Copied festo_gripper.urdf.xacro from mir repo
This file is copied from:
mas-group/mas_industrial_robotics@d9acf9441e666fbb0dbb21b09f3c0918dc3bb302

It's a bit of a hack, the old fingers are left in their just to stop it complaints when using the arm and gripper in simulation.
The new gripper fingers don't move in simulation. URDF only supports tree like structures, and the gripper has linkage.

The simplified model of the festo gripper in the open position, is not just a convex hull. I overlayed the gripper in the open and closed position, projected it onto a new drawing.
Then took a convex hull of the new drawing, but left the middle bit so that moveit wouldn't complain about the object being manipulated as a collision.

Getting the gripper to work in simulation is low on the ToDo list.. getting the arm to work in simulation is probably more important.
Our youbot arm only works in simulation with moveit, but not directly.